requirements.txt
flask
python-dotenv
gunicorn
psycopg2-binary
flask-login
werkzeug
